fix: Correct means sample size
The minimum sample size calculation for means had a result twice as
big as the correct value. Also it asked the user to upload a CSV
with data from a treatment group, which is not necessary. So,
instead, now the user has to input the control ratio.

Besides that, other improvements were added, like:
- File summary
- Three download buttons
- The upload file buttom is at the bottom, to let the parameters
stay together at the top
- Functions to organize the code