feat: Specify configs to avoid st_navbar defaults
If, for some reason, the navbar component can't find the frontend theme object, it uses the config options. Without a config option present it would default to a style that may not be desirable.

Also, specifying the config options helps to keep track of the color pallete used in the web app.